Disable timeout timer when queueing next commit
This prevents problems, where the timeout timer fires before the send
timer. This leads to a reset of the send timer, because the prev_commit
timestamp hasn't been updated since the last invocation of check_timeout(),
and thus, dragontime tries to queue another commit.

Thus, it is necessary to disable to timeout timer while waiting for
sending the next commit frame.